Small renaming improvement
